+++ packages/vue/src/range-slider/range-slider-marker.tsx
@@ -1,11 +1,11 @@
 import { defineComponent, type PropType } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { getValidChildren, type ComponentWithProps } from '../utils'
+import { getValidChildren } from '../utils'
 import { useRangeSliderContext } from './range-slider-context'
 
 export type RangeSliderMarkerProps = HTMLArkProps<'span'> & { value: number }
 
-export const RangeSliderMarker: ComponentWithProps<RangeSliderMarkerProps> = defineComponent({
+export const RangeSliderMarker = defineComponent({
   name: 'RangeSliderMarker',
   props: {
     value: {