+++ packages/vue/src/range-slider/range-slider-thumb.tsx
@@ -1,11 +1,10 @@
 import { defineComponent, type PropType } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import { type ComponentWithProps } from '../utils'
 import { useRangeSliderContext } from './range-slider-context'
 
 export type RangeSliderThumbProps = HTMLArkProps<'div'> & { index: number }
 
-export const RangeSliderThumb: ComponentWithProps<RangeSliderThumbProps> = defineComponent({
+export const RangeSliderThumb = defineComponent({
   name: 'RangeSliderThumb',
   props: {
     index: {